contracts: share the driver failure-kind registry with CLI consumers
The registry holding each failure code's operator-facing meaning and
likely cause lived in the resource-runtime crate, which the d2b CLI does
not depend on. Move it to d2b-contracts, which both sides already reach,
and re-export it from the runtime so every existing import path is
unchanged.

The registry strings and the rendered reference page are byte-identical;
only their home changed. One test still guards the page against drift.
